refactor: separates providers between AppModule and RepositoriesModule
