+++ app/src/main/java/com/omouravictor/ratesnow/framework/di/AppModule.kt
@@ -2,11 +2,6 @@ package com.omouravictor.ratesnow.framework.di
 
 import android.content.Context
 import com.omouravictor.ratesnow.data.local.AppDataBase
-import com.omouravictor.ratesnow.data.network.ApiService
-import com.omouravictor.ratesnow.data.repository.BitCoinRepository
-import com.omouravictor.ratesnow.data.repository.RatesApiRepository
-import com.omouravictor.ratesnow.data.repository.RatesLocalRepository
-import com.omouravictor.ratesnow.data.repository.StocksRepository
 import com.omouravictor.ratesnow.utils.DispatcherProvider
 import dagger.Module
 import dagger.Provides
@@ -28,27 +23,7 @@ object AppModule {
 
     @Singleton
     @Provides
-    fun provideRatesApiRepository(api: ApiService): RatesApiRepository =
-        RatesApiRepository(api)
-
-    @Singleton
-    @Provides
-    fun provideRatesLocalRepository(dataBase: AppDataBase): RatesLocalRepository =
-        RatesLocalRepository(dataBase)
-
-    @Singleton
-    @Provides
-    fun provideStocksRepository(api: ApiService, dataBase: AppDataBase): StocksRepository =
-        StocksRepository(api, dataBase)
-
-    @Singleton
-    @Provides
-    fun provideBitCoinsRepository(api: ApiService, dataBase: AppDataBase): BitCoinRepository =
-        BitCoinRepository(api, dataBase)
-
-    @Singleton
-    @Provides
-    fun provideDispatchers(): DispatcherProvider = object : DispatcherProvider {
+    fun provideAppDispatcherProvider(): DispatcherProvider = object : DispatcherProvider {
         override val main: CoroutineDispatcher
             get() = Dispatchers.Main
         override val io: CoroutineDispatcher
