--- app/src/main/java/com/omouravictor/ratesnow/framework/di/UseCasesModule.kt
@@ -1,12 +1,0 @@
-package com.omouravictor.ratesnow.framework.di
-
-import dagger.Binds
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.android.components.ApplicationComponent
-import javax.inject.Singleton
-
-@Module
-@InstallIn(ApplicationComponent::class)
-interface UseCasesModule {
-}
